refactor(linter): Further break down gitlabci-related tasks
